integrated new changes to irl libre
needed to add a delay to py script to give time for usb to launch
this causes an annoying tho now that oyu have to reboot the machine twice to see changes on usb since the movies.js gets updated after launch of the .html
tried killing two birds with one stone and just launching the .html from the .py after the changes but then the whole script doesnt run because of some driver error AFTER BOOTING the .html. so like when i run the .py from terminal its fine and all just throws an error, but on boot the error prevents the script from running as a whole. >:(